implemented authorization and authentication in user controller

+++ src/main/java/com/masai/controller/UserController.java
@@ -2,12 +2,14 @@ package com.masai.controller;
 
 import java.util.List;
 
-import javax.security.auth.message.AuthException;
 import javax.validation.Valid;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,59 +17,89 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.masai.exception.UserException;
 import com.masai.model.User;
+import com.masai.service.UserHelper;
 import com.masai.service.UserService;
 
 @RestController
 @CrossOrigin("*")
+@RequestMapping("/users")
 public class UserController {
 
 	@Autowired
 	private UserService userService;
+	@Autowired
+	private UserHelper userHelper;
 
-	@PostMapping("/users")
+	private static final Logger logger = LoggerFactory.getLogger(UserController.class);
+
+	@PreAuthorize("hasRole('ADMIN')")
+	@PostMapping("")
 	public ResponseEntity<User> registerUser(@Valid @RequestBody User user) throws UserException {
 		User registeredUser = userService.registerUser(user);
 		return new ResponseEntity<User>(registeredUser, HttpStatus.CREATED);
 	}
 
-	@GetMapping("/users/{email}")
+	@GetMapping("/email/{email}")
 	public ResponseEntity<User> getUserByEmail(@PathVariable("email") String email) throws UserException {
+
+		String loggedInEmail = userHelper.getLoggedInEmail();
+
+		if (!email.equals(loggedInEmail))
+			throw new UserException("Unauthorized to access other user details.");
 
 		User user = userService.getUserByEmail(email);
 
 		return new ResponseEntity<User>(user, HttpStatus.OK);
 	}
 
-	@PutMapping("/users")
+	@PutMapping("")
 	public ResponseEntity<User> updateUser(@Valid @RequestBody User user) throws UserException {
+
+		String loggedInEmail = userHelper.getLoggedInEmail();
+
+		if (!loggedInEmail.equals(user.getEmail()))
+			throw new UserException("Unauthorized to access update user.");
 
 		User existingUser = userService.updateUser(user);
 		return new ResponseEntity<User>(existingUser, HttpStatus.OK);
 	}
 
-	@DeleteMapping("/users/{userId}")
+	@DeleteMapping("/{userId}")
 	public ResponseEntity<User> deleteUser(@PathVariable("userId") Integer userId) throws UserException {
-
-		User user = userService.deleteUser(userId);
-
-		return new ResponseEntity<User>(user, HttpStatus.OK);
-
-	}
-
-	@GetMapping("/getusers/{userId}")
-	public ResponseEntity<User> getUserById(@PathVariable("userId") Integer userId) throws UserException {
+		String loggedInEmail = userHelper.getLoggedInEmail();
 
 		User user = userService.getUserById(userId);
 
+		if (!loggedInEmail.equals(user.getEmail()))
+			throw new UserException("Unauthorized to access delete user with user id : " + userId);
+
+		User duser = userService.deleteUser(userId);
+
+		return new ResponseEntity<User>(duser, HttpStatus.OK);
+
+	}
+
+	@GetMapping("/{userId}")
+	public ResponseEntity<User> getUserById(@PathVariable("userId") Integer userId) throws UserException {
+		User user = null;
+		String loggedInEmail = userHelper.getLoggedInEmail();
+
+		logger.info("logged in email : " + loggedInEmail);
+		user = userService.getUserById(userId);
+
+		if (!loggedInEmail.equals(user.getEmail()))
+			throw new UserException("Unauthorized to access user with user id : " + userId);
+
 		return new ResponseEntity<User>(user, HttpStatus.OK);
 	}
 
-	@GetMapping("/users")
+	@PreAuthorize("hasRole('ADMIN')")
+	@GetMapping("")
 	public ResponseEntity<List<User>> getAllUser() throws UserException {
 		List<User> users = userService.getAllUser();
